feat/Add redirection on signup and logout
close #160
