+++ client/src/context/constants.js
@@ -1,3 +1,4 @@
+const LOGOUT = 'logout';
 const USER = 'user';
 const TITLE = 'title';
 const CONTENT = 'content';
@@ -7,5 +8,5 @@ const CATEGORY_OPTIONS = 'categoryOptions';
 const IS_PUBLIC = 'isPublic';
 
 export {
-  USER, TITLE, CONTENT, CATEGORY, ACTIVE_TAB, CATEGORY_OPTIONS, IS_PUBLIC,
+  LOGOUT, USER, TITLE, CONTENT, CATEGORY, ACTIVE_TAB, CATEGORY_OPTIONS, IS_PUBLIC,
 };
