+++ client/src/context/initialState.js
@@ -1,0 +1,11 @@
+const initialState = {
+  user: {},
+  title: '',
+  content: '',
+  category: '',
+  activeTab: 'write',
+  categoryOptions: [],
+  isPublic: true,
+};
+
+export default initialState;
